+++ src/js/components/RateSelector.vue
